# src/common/api-types/types.go
// Package api_types provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api_types

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error   string  `json:"error"`
	Message string  `json:"message"`
	Stack   *string `json:"stack,omitempty"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status  string `json:"status"`
	Version string `json:"version"`
}

// Location defines model for Location.
type Location struct {
	Crs         *string  `json:"crs,omitempty"`
	FullName    *string  `json:"full_name,omitempty"`
	Stanox      string   `json:"stanox"`
	TiplocCodes []string `json:"tiploc_codes"`
}

// LocationServicesResponse defines model for LocationServicesResponse.
type LocationServicesResponse struct {
	Location Location          `json:"location"`
	Services []ServiceResponse `json:"services"`
}

// NotFoundResponse defines model for NotFoundResponse.
type NotFoundResponse struct {
	Error string `json:"error"`
}

// Operator defines model for Operator.
type Operator struct {
	Code string `json:"code"`
	Name string `json:"name"`
}

// ScheduleLocation defines model for ScheduleLocation.
type ScheduleLocation struct {
	// ActualArrival Actual arrival time from TRUST feed (if available)
	ActualArrival *string `json:"actual_arrival,omitempty"`

	// ActualDeparture Actual departure time from TRUST feed (if available)
	ActualDeparture *string `json:"actual_departure,omitempty"`
	Arrival         *string `json:"arrival,omitempty"`

	// ArrivalLateness Lateness in minutes for arrival (positive = late, negative = early)
	ArrivalLateness *int    `json:"arrival_lateness,omitempty"`
	Departure       *string `json:"departure,omitempty"`

	// DepartureLateness Lateness in minutes for departure (positive = late, negative = early)
	DepartureLateness *int     `json:"departure_lateness,omitempty"`
	Id                int      `json:"id"`
	Location          Location `json:"location"`
	LocationOrder     int      `json:"location_order"`
	LocationType      string   `json:"location_type"`
	Platform          *string  `json:"platform,omitempty"`
	PublicArrival     *string  `json:"public_arrival,omitempty"`
	PublicDeparture   *string  `json:"public_departure,omitempty"`
}

// ServiceResponse defines model for ServiceResponse.
type ServiceResponse struct {
	Headcode          string              `json:"headcode"`
	Id                int                 `json:"id"`
	Locations         []ScheduleLocation  `json:"locations"`
	Operator          *Operator           `json:"operator,omitempty"`
	ScheduleDaysRuns  *string             `json:"schedule_days_runs,omitempty"`
	ScheduleEndDate   *openapi_types.Date `json:"schedule_end_date,omitempty"`
	ScheduleStartDate *openapi_types.Date `json:"schedule_start_date,omitempty"`
	SignallingId      string              `json:"signalling_id"`
	TrainCategory     *string             `json:"train_category,omitempty"`
	TrainStatus       *string             `json:"train_status,omitempty"`
	TrainUid          string              `json:"train_uid"`
}

// GetServiceParams defines parameters for GetService.
type GetServiceParams struct {
	// Headcode The headcode of the services to retrieve
	Headcode string `form:"headcode" json:"headcode"`
}

// GetServicesAtLocationParams defines parameters for GetServicesAtLocation.
type GetServicesAtLocationParams struct {
	// Name The station name to search for trains at
	Name *string `form:"name,omitempty" json:"name,omitempty"`

	// Crs The CRS code of the station to search for trains at
	Crs *string `form:"crs,omitempty" json:"crs,omitempty"`

	// Tiploc The TIPLOC code of the location to search for trains at
	Tiploc *string `form:"tiploc,omitempty" json:"tiploc,omitempty"`

	// Stanox The STANOX code of the location to search for trains at
	Stanox *string `form:"stanox,omitempty" json:"stanox,omitempty"`

	// Date The date to search for services (YYYY-MM-DD format). Defaults to today.
	Date *openapi_types.Date `form:"date,omitempty" json:"date,omitempty"`
}